+++ src/modules/ripv1.c
@@ -39,9 +39,9 @@
  * @param co Pointer to T50 configuration structure.
  * @param size Pointer to packet size (updated by the function).
  */
-void ripv1(const struct config_options *const __restrict__ co, size_t *size)
+void ripv1(const struct config_options *const __restrict__ co, uint32_t *size)
 {
-  size_t length;
+  uint32_t length;
 
   memptr_t buffer;
 
@@ -124,11 +124,11 @@ void ripv1(const struct config_options *const __restrict__ co, size_t *size)
   }
   pseudo->zero     = 0;
   pseudo->protocol = co->ip.protocol;
-  length = (size_t)(buffer.ptr - (void *)udp);
+  length = (uint32_t)(buffer.ptr - (void *)udp);
   pseudo->len      = htons(length);
 
   /* Computing the checksum. */
-  length = (size_t)((void *)(pseudo + 1) - (void *)udp);
+  length = (uint32_t)((void *)(pseudo + 1) - (void *)udp);
   udp->check  = co->bogus_csum ? RANDOM() :
                 cksum(udp, length);
