+++ src/modules/ripv2.c
@@ -39,9 +39,9 @@
  * @param co Pointer to T50 configuration structure.
  * @param size Pointer to packet size (updated by the function).
  */
-void ripv2(const struct config_options *const __restrict__ co, size_t *size)
+void ripv2(const struct config_options *const __restrict__ co, uint32_t *size)
 {
-  size_t greoptlen,     /* GRE options size. */
+  uint32_t greoptlen,     /* GRE options size. */
          length,
          counter;
 
@@ -226,7 +226,7 @@ void ripv2(const struct config_options *const __restrict__ co, size_t *size)
    */
   if (co->rip.auth)
   {
-    size_t size;
+    uint32_t size;
 
     *buffer.word_ptr++ = 0xffff;
     *buffer.word_ptr++ = htons(1);
